refactored cmake scripts
