+++ CMakeLists.txt
@@ -1,15 +1,18 @@
 cmake_minimum_required(VERSION 3.27)
-project(leetcode_cpp)
+
+project(leetcode_cpp CXX)
 
 set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_FLAGS "-Wall -Wextra -Wconversion -Werror -Wpedantic")
 
-add_executable(leetcode_cpp
-        src/Problem129.cpp
-        src/TreeNode.cpp
-        src/Problem32.cpp
-        src/Problem356.cpp
-        src/Problem1702.cpp
-        src/Problem129.cpp
-        src/Problem2370.cpp
-        src/Problem834.cpp
-)
+add_subdirectory(Problem32)
+add_subdirectory(Problem129)
+add_subdirectory(Problem159)
+add_subdirectory(Problem165)
+add_subdirectory(Problem356)
+add_subdirectory(Problem834)
+add_subdirectory(Problem1702)
+add_subdirectory(Problem1915)
+add_subdirectory(Problem2370)
+add_subdirectory(Problem2441)
+add_subdirectory(Problem2997)
