+++ Problem129/CMakeLists.txt
@@ -1,0 +1,2 @@
+include_directories(../common)
+add_executable(problem129 src/Problem129.cpp)
