+++ Problem159/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem159 src/Problem159.cpp)
