+++ Problem165/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem165 src/Problem165.cpp)
