+++ Problem2997/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem2997 src/Problem2997.cpp)
