+++ Problem32/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem32 src/Problem32.cpp)
