+++ Problem356/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem356 src/Problem356.cpp)
