+++ Problem834/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(problem834 src/Problem834.cpp)
